Proxy https in ftp, tests
