+++ src/test/munit/ftp-copy-test-case.xml
@@ -29,6 +29,11 @@
                     <munit:parameter propertyName="config" value="config-proxy-auth"/>
                 </munit:parameters>
             </munit:parameterization>
+            <munit:parameterization name="proxy-https">
+                <munit:parameters>
+                    <munit:parameter propertyName="config" value="config-proxy-https"/>
+                </munit:parameters>
+            </munit:parameterization>
 
         </munit:parameterizations>
     </munit:config>
@@ -53,6 +58,8 @@
                 #[{}]
             </java:args>
         </java:invoke-static>
+        <ftp:delete config-ref="${config}" path="/copy-dir-test"/>
+        <ftp:delete config-ref="${config}" path="/copy-test"/>
     </munit:after-suite>
 
     <ftp:config name="ftp-config">
