+++ src/test/munit/ftp-directory-listener-post-processing-action-test-case.xml
@@ -29,7 +29,11 @@
                     <munit:parameter propertyName="config" value="config-proxy"/>
                 </munit:parameters>
             </munit:parameterization>
-
+            <munit:parameterization name="proxy-https">
+                <munit:parameters>
+                    <munit:parameter propertyName="config" value="config-proxy-https"/>
+                </munit:parameters>
+            </munit:parameterization>
         </munit:parameterizations>
     </munit:config>
