+++ src/test/resources/dockercompose/docker-compose-auth.yml
@@ -1,4 +1,4 @@
-version: "3.9"
+version: "3.8"
 services:
   squidauth:
     image: "robhaswell/squid-authenticated"
