+++ src/test/resources/dockercompose/docker-compose.yml
@@ -1,4 +1,4 @@
-version: "3.9"
+version: "3.8"
 services:
   squid:
     image: "lucacri/squid-ext-conf"
@@ -8,7 +8,7 @@ services:
       - "3129:3129"
     volumes:
       - ./squid.conf:/etc/squid/squid.conf
-
+      - ./myCA.pem:/etc/squid/certs/key.pem
   ftp:
     image: "fauria/vsftpd"
     container_name: "ftp"
