+++ src/test/resources/dockercompose/myCA.pem
@@ -1,0 +1,56 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDAqFd+vpGWOuqY
+hmU1aW2KyIsJyHxk0itDoDrwDh9xuWzQkNwLdNx620njWYqxTtDImj5yrLom3Kvr
+57r7b4kKCvIzNtOBnzIgPdVNdbw/bkPWWiRNZn4AXMWD+uVUUMID/aEg3qMQ3+b7
+J1Tga4cToOcoJwDgg3MlWkXAPEFhL35H0/UN/tayVrgYSG6CRWNu4AIsT4c8EE2a
+EEccJ7j5VsqN7GWp5CQWUUVBFCnM8HiMBzNW2t+L8ePkxuJnaVpI4I6rroBeHe0g
+wLKX7T1kiSvx5aJXkioNa397CzZb+8bWZ00Oa+8hQSepHtUUu8bFdYzcLsOWNjc2
+IzPwa5iFAgMBAAECggEAU3yf7dzHWkLNxqr20N6PCcU9+09WGJq7BOM6pyx3k+C2
+tusPcCeEI/lDW7JB1/Q/pa6xwR/4yz51ghU5ghwjVqbHrknj/hfusOoatSaAHTy7
+SPhnIuN3VtDdBwSCvLcCXJhcatIP+BAPGGdC8L2oG2luwoekaaaND3K/Jz71Vhdt
+LCJx+1+XXPlJxKUNtxMCXNC9gtnq8m4tEAt/v7EyjMXF/4vkulwKV/c1iopvvUL9
+0Eyf70uMKja5WqtCfwN7vuF9OUeWOsKDtEfzYMGYbLbY7LBqgxKe1aPvglUN55lE
+SvWTFM5VtDJM7p4pSgojeFBMWbLc8EoGowKsS+NQAQKBgQD2cwYHxcjcKHniyTjk
+shsfmZSo0wJZ53kCuqrEsTan09X3djMM0qn+OujP7CHdhOmIjqsnACWxxWFm3JDB
+LCY/qdXpa/n6WzTUqB0otJ+r1Bevn+K15RZCFSpQ4lDhfe5RXLM/SJj7qEoGgNvA
+97ICtm5lI26BL2GNaj9/yygewQKBgQDIH6kqmPZ+rg6sp+oEumFrEGYU9CHt53lx
+QKvkrbWgPIEWisx9j9QjXn6xBnzzOZR3iCeMTamJmGLhi2o4gN57YAo/eRBBVsVX
+b7xrqpPeKcLsECkUE+XlAsIStGDD83o96RnAFgAkoWunR1CQgYhGILvT1i4Zacvy
+g0KcN5VuxQKBgQCzbi1uzUafW+zZ/A0rHK14xVWh0Pd+Q3QsU4HvcOLNMAMxh2JF
+rL9NtNA6Sb6iTJnG7bCY1ZecJYvz2xSdQCqc4Wc36oB3EZ6WUcvhDhjOfr28BVmH
+e0JIXKaReyTN9+TiYe6zXYTcfLtgnsLgbLazGOWpRzG2ZqvMzRVsUppkQQKBgGpM
+fvPUeiIB3GOTcXd/dJ2FAWQh5R/UMyqX+nYmWvI5ESxjH4mWa006NTtcweQD8run
+Q7J/0gz3N5asEgjfUs3fzrG2YY8DQyFVhZT+7zoznoj7cIZsp1GErId47mlHG+hk
+t7vTM9JBN8iL1dxcIYF2g65a9vOiwaUYWbpWv/YtAoGBAIiipsoJpF0vLOMsfm8L
+fY204kYI3tXZUNIWvJu550v9GnIXzujb7A3oxA+EPfUg4EBITLz5okigwC2De/4s
+xSvClNSoPLChkiJiy4Ykh84vwmEU3gwdflsXQ8kOzpjcXTWmm+D6FtshWgz52Mu9
+k8bBpu6sD1qA5kgT8vEFVNEM
+-----END PRIVATE KEY-----
+-----BEGIN CERTIFICATE-----
+MIIEyjCCA7KgAwIBAgIJAI5/xkNDhGGgMA0GCSqGSIb3DQEBCwUAMIGdMQswCQYD
+VQQGEwJBUjEUMBIGA1UECAwLQnVlbm9zQWlyZXMxFDASBgNVBAcMC0J1ZW5vc0Fp
+cmVzMREwDwYDVQQKDAhNdWxlc29mdDEXMBUGA1UECwwOQ29yZUNvbm5lY3RvcnMx
+EzARBgNVBAMMCkNvbW1vbk5hbWUxITAfBgkqhkiG9w0BCQEWEmZuZXJ5QG11bGVz
+b2Z0LmNvbTAeFw0yMjAyMjUxNDE4NDVaFw0yMzAyMjUxNDE4NDVaMIGdMQswCQYD
+VQQGEwJBUjEUMBIGA1UECAwLQnVlbm9zQWlyZXMxFDASBgNVBAcMC0J1ZW5vc0Fp
+cmVzMREwDwYDVQQKDAhNdWxlc29mdDEXMBUGA1UECwwOQ29yZUNvbm5lY3RvcnMx
+EzARBgNVBAMMCkNvbW1vbk5hbWUxITAfBgkqhkiG9w0BCQEWEmZuZXJ5QG11bGVz
+b2Z0LmNvbTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAMCoV36+kZY6
+6piGZTVpbYrIiwnIfGTSK0OgOvAOH3G5bNCQ3At03HrbSeNZirFO0MiaPnKsuibc
+q+vnuvtviQoK8jM204GfMiA91U11vD9uQ9ZaJE1mfgBcxYP65VRQwgP9oSDeoxDf
+5vsnVOBrhxOg5ygnAOCDcyVaRcA8QWEvfkfT9Q3+1rJWuBhIboJFY27gAixPhzwQ
+TZoQRxwnuPlWyo3sZankJBZRRUEUKczweIwHM1ba34vx4+TG4mdpWkjgjquugF4d
+7SDAspftPWSJK/HloleSKg1rf3sLNlv7xtZnTQ5r7yFBJ6ke1RS7xsV1jNwuw5Y2
+NzYjM/BrmIUCAwEAAaOCAQkwggEFMA8GA1UdEwEB/wQFMAMBAf8wHQYDVR0OBBYE
+FBR+zRHKEytBMWqo8Bg1ex1qiMi1MIHSBgNVHSMEgcowgceAFBR+zRHKEytBMWqo
+8Bg1ex1qiMi1oYGjpIGgMIGdMQswCQYDVQQGEwJBUjEUMBIGA1UECAwLQnVlbm9z
+QWlyZXMxFDASBgNVBAcMC0J1ZW5vc0FpcmVzMREwDwYDVQQKDAhNdWxlc29mdDEX
+MBUGA1UECwwOQ29yZUNvbm5lY3RvcnMxEzARBgNVBAMMCkNvbW1vbk5hbWUxITAf
+BgkqhkiG9w0BCQEWEmZuZXJ5QG11bGVzb2Z0LmNvbYIJAI5/xkNDhGGgMA0GCSqG
+SIb3DQEBCwUAA4IBAQA4Jeckmz2/nEQ8V5n23vIAaHoWMMq+R8d/1XGxNYUMaTC3
+AD9FsI1NFzj0R96J6kREKsD4UP+gn+sXxzr8H/AuKiGLUKb7O7CzkMgDnrxFvRPZ
+3FDFINcWlHdRtV5iu2/2Jkof+8hipFa3jzqmPKHexZsJMGNq+L0QvMRh2eM6Qp0N
+ZGnvS79e63LF6MfuoQyP/LXl8GKPP4OF8w9VeLkh/fsR+zqmVPuZAfUKKlZQn0r+
+Ug4uDnnxjk05kFzwR5ERDDtcrqdH4ADNFZET3XgPdCqgwyGQLQn4jkWuA2+tkK+D
+Zq1amp66CZGqMvYnwDW4Q/QnvLQ2tYshmiIjaLwY
+-----END CERTIFICATE-----
