+++ src/test/resources/dockercompose/squid.conf
@@ -7,10 +7,13 @@ http_access allow localhost manager
 http_access deny manager
 
 acl ftp proto FTP
+acl ftps proto FTPS
 
 http_access allow ftp
+http_access allow ftps
 
 http_access allow localhost
 http_access allow all
 
 http_port 3128
+https_port 3129 tls-cert=/etc/squid/certs/key.pem
